+++ src/app/pages/profile/profile.component.ts
@@ -5,9 +5,10 @@ import { Observable } from 'rxjs';
 
 import { UserService } from '@services/users.service';
 import { AuditService } from '@services/audit.service';
-import { IUser } from 'src/app/core/models/user';
-import { IAuditItem } from 'src/app/core/models/audit';
-import { BaseType } from 'src/app/core/models/base';
+import { EventService } from '@services/events.service';
+import { BaseType, IBase } from '@models/base';
+import { IUser } from '@models/user';
+import { IAuditItem } from '@models/audit';
 
 
 @Component({
@@ -19,18 +20,22 @@ export class ProfileComponent implements OnInit {
 
   public userData$: Observable<IUser>;
   public auditItems: IAuditItem[] = [];
+  public createdEventItems: IBase[] = [];
   public baseTypeEntity: BaseType = BaseType.ENTITY;
   public baseTypeAudit: BaseType = BaseType.AUDIT;
+  public baseTypeEvent: BaseType = BaseType.EVENT;
 
   constructor(
     public auth: AngularFireAuth,
     private auditSrv: AuditService,
+    private eventSrv: EventService,
     private userSrv: UserService
   ) { }
 
   ngOnInit(): void {
     this.auth.user.subscribe((usr) => {
       const uidUser = usr.uid;
+      this.getCreatedEvents(uidUser);
       this.userData$ = this.userSrv.getOneUser(uidUser);
       this.getAudit(usr.uid);
     });
@@ -42,4 +47,12 @@ export class ProfileComponent implements OnInit {
       this.auditItems = auditItems;
     });
   }
+
+  getCreatedEvents(uidUser: string): void {
+    this.eventSrv.getAllEventsByUser(uidUser)
+    .subscribe( (eventItems: IBase[]) => {
+      this.createdEventItems = eventItems;
+    });
+  }
+
 }