+++ src/environments/environment.prod.ts
@@ -1,6 +1,6 @@
 export const environment = {
   production: true,
-  appVersion: '0.1.02',
+  appVersion: '0.1.2',
   baseUrl: 'https://info-rincon.web.app/',
   firebase: {
     apiKey: 'AIzaSyChFHMB9Kj4sEBdPKLRWu-JFVMg7gketuM',